Changed all references to latest ESRI 3.5 from 3.2 and made everything work, passes core tests.
it now checks the api is >= 3.4 as 3.3 and below dont work now (not that 3.2 was working before)
changed to use new dojo methods, and based the esri.html off leaflet.html for a fresh start
implemented some missing functionality in ESRI like tilelayer, and used the ESRI base layers for the map control and all default layers.
simplified the esri.html example so it only has one esri specific (dojo) call.